[refactor] investment 기획서 및 테이블 명세서에 맞게 refactoring

+++ src/main/java/com/finlearn/simulationservice/application/investment/dto/response/StockPriceResponse.java
@@ -1,9 +1,7 @@
 package com.finlearn.simulationservice.application.investment.dto.response;
 
-import java.math.BigDecimal;
-
 public record StockPriceResponse(
         String stockCode,
-        BigDecimal currentPrice
+        long currentPrice
 ) {
 }
